feat: Add CreateTaskDto and TaskMapper
This commit adds the `CreateTaskDto` class to the `create-task.dto.ts` file, which represents the data transfer object for creating a task. It also adds the `TaskMapper` class to the `task.mapper.ts` file, which includes a static method `taskEntityFromObject` for mapping task data from an object to a `TaskEntity` instance.

These changes are necessary to support the creation of tasks in the application.