feat: Add getTasksByProjectId method to TaskDataSource
This commit adds the `getTasksByProjectId` method to the `TaskDataSource` class in the `task.datasource.impl.ts` file. The method retrieves tasks by project ID and returns an array of `TaskEntity` objects. This implementation enhances the functionality of the application by providing a way to fetch tasks specific to a project.