feat: Add status field to UpdateTaskDto
This commit adds a new field `status` to the `UpdateTaskDto` class in the `update-task.dto.ts` file. The `status` field allows for updating the status of a task. This enhancement improves the flexibility and functionality of the application by providing the ability to modify the status of a task during an update operation.