feat: Add createdAt and updatedAt fields to TaskEntity
This commit adds the `createdAt` and `updatedAt` fields to the `TaskEntity` class in the `task.entity.ts` file. These fields represent the creation and last update timestamps of a task. This enhancement improves the functionality of the application by providing the ability to track the creation and modification times of tasks.