feat: Add manager field to project model
This commit adds a `manager` field to the `IProject` interface and `projectSchema` in the `project.model.ts` file. The `manager` field is a reference to the `User` model and is required for each project. This enhancement allows for better organization and management of projects by assigning a manager to each project.